+++ CMakeLists.txt
@@ -69,7 +69,6 @@ add_library(utilities_lib STATIC
     src/simulation/GeometryProcessing.cpp
     src/simulation/LinearAlgebra.cpp
     src/simulation/Validation.cpp
-    src/simulation/utilities/sim_structs.cpp
 )
 target_include_directories(utilities_lib PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/include")
 target_link_libraries(utilities_lib PRIVATE CGAL::Eigen3_support Boost::boost Boost::filesystem assimp::assimp)
