+++ include/IO.h
@@ -16,19 +16,16 @@
 #include <assimp/scene.h>
 #include <assimp/postprocess.h>
 
-#include <utilities/sim_structs.h>
-
 const boost::filesystem::path PROJECT_PATH_IO = PROJECT_SOURCE_DIR;
-
 
 void loadMeshVertices(std::string filepath, Eigen::MatrixXd& vertices);
 
 void loadMeshFaces(std::string filepath, Eigen::MatrixXi& faces);
 
-std::pair<Eigen::MatrixXd, std::vector<int64_t>> get_mesh_data(
-    std::unordered_map<int, Mesh_UV_Struct> mesh_dict,
-    int mesh_id
-);
+// std::pair<Eigen::MatrixXd, std::vector<int64_t>> get_mesh_data(
+//     std::unordered_map<int, Mesh_UV_Struct> mesh_dict,
+//     int mesh_id
+// );
 
 
 // We need do define it in the header file or otherwise the template specialization will not be available at link time
