+++ src/simulation/Validation.cpp
@@ -9,14 +9,14 @@
 
 #include <Validation.h>
 
-bool Validation::are_all_valid(const std::vector<VertexData>& vertex_data) {
-    for (const VertexData& data : vertex_data) {
-        if (!data.valid) {
-            return false;
-        }
-    }
-    return true;
-}
+// bool Validation::are_all_valid(const std::vector<VertexData>& vertex_data) {
+//     for (const VertexData& data : vertex_data) {
+//         if (!data.valid) {
+//             return false;
+//         }
+//     }
+//     return true;
+// }
 
 bool Validation::checkForInvalidValues(
     const Eigen::Matrix<double, Eigen::Dynamic, 2> matrix
@@ -31,13 +31,13 @@ bool Validation::checkForInvalidValues(
     return false;
 }
 
-void Validation::error_unvalid_vertices(
-    std::vector<VertexData> vertex_data
-){
-    if (!are_all_valid(vertex_data)) {
-        throw std::runtime_error("There are still particles outside the mesh");
-    }
-}
+// void Validation::error_unvalid_vertices(
+//     std::vector<VertexData> vertex_data
+// ){
+//     if (!are_all_valid(vertex_data)) {
+//         throw std::runtime_error("There are still particles outside the mesh");
+//     }
+// }
 
 void Validation::error_invalid_values(
     Eigen::Matrix<double, Eigen::Dynamic, 2> r_UV_new
