--- src/simulation/utilities/sim_structs.cpp
@@ -1,6 +1,0 @@
-// author: @Jan-Piotraschke
-// date: 2023-04-14
-// license: Apache License 2.0
-// version: 0.1.0
-
-#include <utilities/sim_structs.h>
